+++ charts/traceability-foss/CHANGELOG.md
@@ -12,6 +12,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Migrated from irs-edc-consumer to tractusx-connector
 - Added BASF instance to int environment
 
+## [1.3.32] - 2024-03-04
+### No changes
+
 ## [1.3.31] - 2024-03-04
 ### No changes